added profile pic change button

+++ backend/controllers/post.js
@@ -84,6 +84,11 @@ exports.deletePost = async (req, res) => {
     if (!post) {
       return res.status(404).json({ message: 'Post not found' });
     }
+
+    if (post.userId.toString() !== req.user._id.toString()) {
+      return res.status(401).json({ message: 'Not authorized to delete post' });
+    }
+
     await post.deleteOne();
     res.status(200).json({ message: 'Post deleted successfully' });
   } catch (error) {
@@ -162,6 +167,7 @@ exports.getPostsSortedByMostComments = async (req, res) => {
     const posts = await PostSchema.find()
       .populate('userId')
       .populate('comments')
+      .populate('likes')
       .populate({
         path: 'comments',
         populate: {
@@ -170,9 +176,35 @@ exports.getPostsSortedByMostComments = async (req, res) => {
       });
 
     posts.sort((a, b) => {
-      return b.comments.length - a.comments.length;
+      return (
+        b.comments.length +
+        (b.likes ? b.likes.length : 0) -
+        (a.comments.length + (a.likes ? a.likes.length : 0))
+      );
     });
     res.status(200).json(posts);
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
+};
+
+// add a like to a post
+exports.handleLike = async (req, res) => {
+  try {
+    const post = await PostSchema.findById(req.params.postId);
+    if (!post) {
+      return res.status(404).json({ message: 'Post not found' });
+    }
+    if (post.likes.includes(req.user._id)) {
+      // remove req.user._id from post.likes
+      post.likes = post.likes.filter(
+        (like) => like.toString() !== req.user._id.toString()
+      );
+    } else {
+      post.likes.push(req.user._id);
+    }
+    await post.save();
+    res.status(200).json(post.likes);
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
